feat: 지도 모든 Entity zIndex, globalZIndex 추가

+++ ios/reactNativeNMap/RNNaverMapPathOverlayManager.m
@@ -39,8 +39,8 @@ RCT_EXPORT_VIEW_PROPERTY(outlineColor, UIColor)
 RCT_EXPORT_VIEW_PROPERTY(passedOutlineColor, UIColor)
 RCT_EXPORT_VIEW_PROPERTY(pattern, NSString)
 RCT_EXPORT_VIEW_PROPERTY(patternInterval, CGFloat)
-RCT_EXPORT_VIEW_PROPERTY(zIndex, CGFloat)
-RCT_EXPORT_VIEW_PROPERTY(globalZIndex, CGFloat)
+RCT_EXPORT_VIEW_PROPERTY(zIndex, NSInteger)
+RCT_EXPORT_VIEW_PROPERTY(globalZIndex, NSInteger)
 
 
 @end
